Time: 5 ms (19.92%), Space: 5.8 MB (96.18%) - LeetHub

+++ 50-powx-n/50-powx-n.cpp
@@ -2,17 +2,19 @@ class Solution {
 public:
     double myPow(double x, int n) {
        double ans=1.0;
-        long long nn=n;
-        if(nn<0)nn=-1*nn;
+    long nn=n;
+        if(nn<0)nn=(-1)*nn;
         while(nn){
-            if(nn%2==0){
-                x*=x;nn/=2;}
+          if(nn%2==0){
+              x*=x;
+              nn/=2;
+          }
             else{
                 ans*=x;
                 nn-=1;
             }
         }
-        if(n<0)ans=double(1.0)/double(ans);
+        if(n<0)ans=(double)(1.0)/(double)(ans);
         return ans;
     }  //TC=O(logN)
 };
